requirements.txt
django-sortedm2m~=3.1.1
django-reversion~=4.0.1
django-taggit~=2.1.0
# For prototype
netjsonconfig @ https://github.com/Aryamanz29/netjsonconfig/tarball/zerotier-prototype
django-x509 @ https://github.com/openwisp/django-x509/tarball/master
django-loci @ https://github.com/openwisp/django-loci/tarball/master
django-flat-json-widget @ https://github.com/openwisp/django-flat-json-widget/tarball/master
openwisp-users @ https://github.com/openwisp/openwisp-users/tarball/master
openwisp-notifications @ https://github.com/openwisp/openwisp-notifications/tarball/master
openwisp-ipam @ https://github.com/openwisp/openwisp-ipam/tarball/master
djangorestframework-gis~=0.18.0
paramiko[ed25519]~=2.10.3
scp~=0.14.2
celery~=5.2.3
django-cache-memoize~=0.1
shortuuid~=1.0.1
netaddr~=0.8.0
django-import-export~=2.8.0
